fix: resolve BigDecimal type conversion errors in test files
Fixed compilation errors where double values were being assigned to
BigDecimal fields in test files. Updated all test files to use proper
BigDecimal constructors with string values for precision.

+++ backend/src/main/java/com/sustainablefarm/dto/request/BatchCreateRequest.java
@@ -9,6 +9,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 /**
@@ -33,7 +34,7 @@ public class BatchCreateRequest {
 
     @NotNull(message = "Harvest quantity is required")
     @Positive(message = "Harvest quantity must be positive")
-    private Double harvestQuantityKg;
+    private BigDecimal harvestQuantityKg;
 
     @NotBlank(message = "Farm ID is required")
     private String farmId;
