+++ backend/src/main/java/com/sustainablefarm/dto/request/DryingRunCreateRequest.java
@@ -10,6 +10,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -31,26 +32,26 @@ public class DryingRunCreateRequest {
 
     @NotNull(message = "Duration is required")
     @Positive(message = "Duration must be positive")
-    private Double durationHours;
+    private BigDecimal durationHours;
 
     @NotNull(message = "Target temperature is required")
-    private Double targetTemperatureC;
+    private BigDecimal targetTemperatureC;
 
     @NotNull(message = "Actual temperature is required")
-    private Double actualTemperatureC;
+    private BigDecimal actualTemperatureC;
 
     @NotNull(message = "Start moisture is required")
     @Positive(message = "Start moisture must be positive")
-    private Double startMoisturePct;
+    private BigDecimal startMoisturePct;
 
     @NotNull(message = "End moisture is required")
     @DecimalMin(value = "6.0", message = "End moisture must be at least 6% for EU compliance")
     @DecimalMax(value = "18.0", message = "End moisture must be at most 18% for EU compliance")
-    private Double endMoisturePct;
+    private BigDecimal endMoisturePct;
 
     @NotNull(message = "Energy usage is required")
     @PositiveOrZero(message = "Energy usage cannot be negative")
-    private Double energyUsageKwh;
+    private BigDecimal energyUsageKwh;
 
     @NotNull(message = "Start time is required")
     private LocalDateTime startTime;
